move SafeLineLoader
